+++ dubbo-samples-version/src/main/java/org/apache/dubbo/samples/version/EmbeddedZooKeeper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.dubbo.samples.version;
+package org.apache.dubbo.samples.version;
 
 import java.io.File;
 import java.lang.reflect.Method;
