+++ dubbo-samples-version/src/main/java/org/apache/dubbo/samples/version/VersionConsumer.java
@@ -17,14 +17,27 @@
  *
  */
 
-package com.alibaba.dubbo.samples.version.impl;
+package org.apache.dubbo.samples.version;
 
-import com.alibaba.dubbo.samples.version.api.VersionService;
+import org.apache.dubbo.samples.version.api.VersionService;
 
-public class VersionServiceImpl implements VersionService {
+import org.springframework.context.support.ClassPathXmlApplicationContext;
 
-    public String sayHello(String name) {
-        return "hello, " + name;
+/**
+ * VersionConsumer
+ */
+public class VersionConsumer {
+
+    public static void main(String[] args) throws Exception {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/version-consumer.xml"});
+        context.start();
+        VersionService versionService = (VersionService) context.getBean("versionService");
+        for (int i = 0; i < 10000; i++) {
+            String hello = versionService.sayHello("world");
+            System.out.println(hello);
+            Thread.sleep(2000);
+        }
+        System.in.read();
     }
 
 }
