+++ dubbo-samples-version/src/main/java/org/apache/dubbo/samples/version/impl/VersionServiceImpl.java
@@ -17,27 +17,15 @@
  *
  */
 
-package com.alibaba.dubbo.samples.version;
+package org.apache.dubbo.samples.version.impl;
 
-import com.alibaba.dubbo.samples.version.api.VersionService;
+import org.apache.dubbo.samples.version.api.VersionService;
 
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+public class VersionServiceImpl implements VersionService {
 
-/**
- * VersionConsumer
- */
-public class VersionConsumer {
-
-    public static void main(String[] args) throws Exception {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/version-consumer.xml"});
-        context.start();
-        VersionService versionService = (VersionService) context.getBean("versionService");
-        for (int i = 0; i < 10000; i++) {
-            String hello = versionService.sayHello("world");
-            System.out.println(hello);
-            Thread.sleep(2000);
-        }
-        System.in.read();
+    @Override
+    public String sayHello(String name) {
+        return "hello, " + name;
     }
 
 }
