+++ dubbo-samples-version/src/main/resources/spring/version-consumer.xml
@@ -28,6 +28,6 @@
 
     <dubbo:registry address="zookeeper://127.0.0.1:2181"/>
 
-    <dubbo:reference id="versionService" interface="com.alibaba.dubbo.samples.version.api.VersionService" version="*"/>
+    <dubbo:reference id="versionService" interface="org.apache.dubbo.samples.version.api.VersionService" version="*"/>
 
 </beans>
