+++ dubbo-samples-version/src/main/resources/spring/version-provider.xml
@@ -30,10 +30,10 @@
 
     <dubbo:protocol name="dubbo" port="20880"/>
 
-    <bean id="versionService" class="com.alibaba.dubbo.samples.version.impl.VersionServiceImpl"/>
+    <bean id="versionService" class="org.apache.dubbo.samples.version.impl.VersionServiceImpl"/>
 
 
-    <dubbo:service interface="com.alibaba.dubbo.samples.version.api.VersionService" version="1.0.0"
+    <dubbo:service interface="org.apache.dubbo.samples.version.api.VersionService" version="1.0.0"
                    ref="versionService"/>
 
 </beans>
